+++ src/osrs_toolkit/__init__.py
@@ -1,3 +1,3 @@
 """OSRS Toolkit package."""
 
-__version__ = "1.3.1"
+__version__ = "1.3.2"